requirements.txt
selenium
python-slugify
pandas